+++ components/settings/AgeSlider.vue
