+++ components/settings/DistanceSlider.vue
