+++ pages/app/chat/test2.vue
@@ -25,9 +25,9 @@
 </template>
 
 <script>
-import VMessageTimestamp from '~/components/VMessageTimestamp'
-import VMessageDate from '~/components/VMessageDate'
-import VSendFooter from '~/components/VSendFooter'
+import VMessageTimestamp from '~/components/chat/VMessageTimestamp'
+import VMessageDate from '~/components/chat/VMessageDate'
+import VSendFooter from '~/components/chat/VSendFooter'
 export default {
     layout: 'chat',
     components: {
